chore: directory restructure
