+++ apps/client/src/views/cuesheet/cuesheet-table/cuesheet-table-settings/CuesheetShareModal.tsx
@@ -6,7 +6,7 @@ import RotatedLink from '../../../../common/components/icons/RotatedLink';
 import Modal from '../../../../common/components/modal/Modal';
 import useInfo from '../../../../common/hooks-query/useInfo';
 import useUrlPresets from '../../../../common/hooks-query/useUrlPresets';
-import GenerateLinkFormExport from '../../../../features/app-settings/panel/feature-panel/GenerateLinkFormExport';
+import GenerateLinkFormExport from '../../../../features/sharing/GenerateLinkFormExport';
 
 function CuesheetShareModal() {
   const { data: infoData } = useInfo();
@@ -26,6 +26,7 @@ function CuesheetShareModal() {
         isOpen={isOpen}
         onClose={handler.close}
         title='Share cuesheet view'
+        showBackdrop
         showCloseButton
         bodyElements={
           showModalContent ? <GenerateLinkFormExport lockedPath={{ value: 'cuesheet', label: 'Cuesheet' }} /> : null
